+++ lib/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - add standalone version of the library
+### Changed
+- remove chart `onInit` prop
 
 ## [0.4.1] - 2025-06-06
 ### Changed